calc: extract .py command to its own file
